# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AP Automation API",
    "description": "API for accounts payable automation with ERP-neutral extraction and configurable connectors",
    "version": "1.0.0"
  },
  "paths": {
...
    "/dashboard": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Mission Control Dashboard",
        "description": "Returns the complete Mission Control dashboard data including pipeline state, packages, and statistics.",
        "operationId": "get_mission_control_dashboard_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Period filter (e.g., '2025-11' or 'November 2025')",
              "title": "Period"
            },
            "description": "Period filter (e.g., '2025-11' or 'November 2025')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MissionControlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard/insights/{role}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Role-Based Insights",
        "description": "Returns drill-down insights for a specific stakeholder role.",
        "operationId": "get_insights_dashboard_insights__role__get",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/StakeholderRole",
              "description": "Stakeholder role"
            },
            "description": "Stakeholder role"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrilldownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/dashboard/packages": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "List Packages",
        "description": "Returns list of feedlot packages with optional filtering.",
        "operationId": "list_packages_dashboard_packages_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Period filter (e.g., '2025-11')",
              "title": "Period"
            },
            "description": "Period filter (e.g., '2025-11')"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PackageStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by feedlot name or owner",
              "title": "Search"
            },
            "description": "Search by feedlot name or owner"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Maximum results",
              "default": 50,
              "title": "Limit"
            },
            "description": "Maximum results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/models__api_responses__PackageSummary"
                  },
                  "title": "Response List Packages Dashboard Packages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/packages/{package_id}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Package Details",
        "description": "Returns complete package details including all invoices.",
        "operationId": "get_package_dashboard_packages__package_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Package identifier",
              "title": "Package Id"
            },
            "description": "Package identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/packages/{package_id}/invoices": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "List Package Invoices",
        "description": "Returns all invoices in a package.",
        "operationId": "list_package_invoices_dashboard_packages__package_id__invoices_get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Package identifier",
              "title": "Package Id"
            },
            "description": "Package identifier"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvoiceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by invoice status",
              "title": "Status"
            },
            "description": "Filter by invoice status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceSummary"
                  },
                  "title": "Response List Package Invoices Dashboard Packages  Package Id  Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/packages/{package_id}/invoices/{invoice_id}": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Invoice Details",
        "description": "Returns complete invoice details including line items, validation, and reconciliation.",
        "operationId": "get_package_invoice_dashboard_packages__package_id__invoices__invoice_id__get",
        "parameters": [
          {
            "name": "package_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Package identifier",
              "title": "Package Id"
            },
            "description": "Package identifier"
          },
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Invoice identifier",
              "title": "Invoice Id"
            },
            "description": "Invoice identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/dashboard/drilldown": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Get Drilldown Data",
        "description": "Returns filtered drilldown data with recommended focus target.",
        "operationId": "get_drilldown_dashboard_drilldown_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Drilldown type: stage, reason, check, role",
              "title": "Type"
            },
            "description": "Drilldown type: stage, reason, check, role"
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Specific ID to filter by",
              "title": "Id"
            },
            "description": "Specific ID to filter by"
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Period filter (e.g., '2025-11')",
              "title": "Period"
            },
            "description": "Period filter (e.g., '2025-11')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrilldownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissionControlResponse": {
        "properties": {
          "period": {
            "type": "string",
            "title": "Period",
            "description": "Current period (e.g., 'November 2025')"
          },
          "period_start": {
            "type": "string",
            "format": "date",
            "title": "Period Start",
            "description": "Period start date"
          },
          "period_end": {
            "type": "string",
            "format": "date",
            "title": "Period End",
            "description": "Period end date"
          },
          "last_sync": {
            "type": "string",
            "title": "Last Sync",
            "description": "Relative time since last sync"
          },
          "last_sync_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Sync At",
            "description": "Absolute sync timestamp"
          },
          "pipeline": {
            "$ref": "#/components/schemas/PipelineSnapshot"
          },
          "human_review": {
            "$ref": "#/components/schemas/HumanReviewSummary"
          },
          "packages": {
            "items": {
              "$ref": "#/components/schemas/models__api_responses__PackageSummary"
            },
            "type": "array",
            "title": "Packages"
          },
          "today_stats": {
            "$ref": "#/components/schemas/TodayStats"
          },
          "insights_available": {
            "items": {
              "$ref": "#/components/schemas/StakeholderRole"
            },
            "type": "array",
            "title": "Insights Available",
            "description": "Available insight views"
          }
        },
        "type": "object",
        "required": [
          "period",
          "period_start",
          "period_end",
          "last_sync",
          "last_sync_at",
          "pipeline",
          "human_review",
          "today_stats"
        ],
        "title": "MissionControlResponse",
        "description": "Complete Mission Control dashboard response.\n\nThis is the primary API response for the main dashboard view.\nIt includes:\n- Period information\n- Pipeline state\n- Human review queue\n- Package list\n- Today's statistics\n- Role-based insights summary"
      },
...
      "StakeholderRole": {
        "type": "string",
        "enum": [
          "CFO",
          "COO",
          "CIO",
          "Accounting"
        ],
        "title": "StakeholderRole",
        "description": "Stakeholder roles for role-based views."
      },
...
    }
  }
}